data_fetcher renamed to handler, code refactor, common code moved to handler
