+++ web_app/1_potential_areas/app/handler.py
@@ -1,5 +1,9 @@
 import pandas as pd
 import numpy as np
+import pickle
+import streamlit as st
+
+LOOKER_URL = "https://lookerstudio.google.com/embed/reporting/eaab71cb-575f-4f7c-b9ac-97942e43d017/page/r2W2D"
 
 class DataFetcher:
     def __init__(self, parquet_path):
@@ -16,3 +20,20 @@ class DataFetcher:
         if row['distance'] > tolerance:
             return None
         return row.to_dict()
+
+
+# Load the models 
+def load_model(model_path):
+    with open(model_path, "rb") as f:
+        model = pickle.load(f)
+    return model
+
+# Read parquet file
+@st.cache_data
+def read_parquet(filename):
+    df = pd.read_parquet(
+        filename,
+        engine="pyarrow",
+        dtype_backend="pyarrow",
+    )
+    return df
